+++ base/draw.h
@@ -49,8 +49,8 @@ void openclose(double x, double y1, double y2, double ebarlen, int xy);
 void putlegend(int i,		/* which set */
 	        int d,		/* flag, 1 = no draw, just compute min/max
 				 * for bounding box */
-	        int xlen,	/* length of legend */
-	        int ylen,	/* distance between entries */
+                double xlen,	/* length of legend */
+                double ylen,	/* distance between entries */
 	        double size,	/* symbol size */
 	        double x,	/* location x */
 	        double y,	/* location y */
@@ -66,8 +66,8 @@ void putlegend(int i,		/* which set */
 void putbarlegend(int i,	/* which set */
 	        int d,		/* flag, 1 = no draw, just compute min/max
 				 * for bounding box */
-	        int xlen,	/* length of legend */
-	        int ylen,	/* distance between entries */
+                double xlen,	/* length of legend */
+                double ylen,	/* distance between entries */
 	        double size,	/* symbol size */
 	        double x,	/* location x */
 	        double y,	/* location y */
